requirements.txt
langchain>=0.3.7
langchain-community>=0.3.7
langgraph>=0.2.39
groq>=0.11.0
chromadb>=0.5.5
sentence-transformers>=3.1.1
python-dotenv>=1.0.1
pypdf>=5.0.0
streamlit>=1.38.0
tqdm>=4.66.4
